tools/birdsong-mix/requirements.txt
pydub>=0.25.1
requests>=2.31.0
PyYAML>=6.0.1
